✨ feat: current-utterance focus — dim past rows during playback
While a run is in flight, dim non-current log rows to 0.65 so the eye
settles on the line being revealed (chat-log analogue of a VN current-line
stage). The latest .agentOutput stays full; the in-flight streaming row is
inherently full (outside the dimmed ForEach). Focus releases to full
opacity when the run ends/pauses so the completed log reads normally.
Pure focusedOpacity helper + pastFocusOpacity constant.

Playback-display tests (focus opacity + reading-pause display length) moved
to a sibling extension file (SimulationViewModelStreamingTests+Focus) to
keep the suite under the type_body_length cap; makeSUT widened to internal
for the sibling per testing.md.
