+++ Pastura/Pastura/Resources/Localizable.xcstrings
@@ -6468,7 +6468,6 @@
       }
     },
     "~%lld inferences" : {
-      "extractionState" : "stale",
       "localizations" : {
         "ja" : {
           "stringUnit" : {
